Poultry_API/requirements.txt
fastapi
uvicorn
python-multipart
numpy
tensorflow
pillow
matplotlib
google-generativeai